dt-bindings: irqchip: ti,pruss-intc: add interrupt cell for AM18xx
The PRUSS interrupt controller on TI AM18xx/OMAP-L138 SoCs has an
event selector that is not found on other PRUSSes. For events 0 to 31,
the actual event depends on the setting of PRUSSEVTSEL [1], so we need
an additional cell to fully describe the requested interrupt.

This changes the binding to allow 3 or 4 cells. When #interupt-cells is
set to 4, then the ti,da830-cfgchip property is required. This provides
a link to the syscon register where PRUSSEVTSEL lives.

[1]: https://processors.wiki.ti.com/index.php/PRU_Subsystem_Differences_Between_OMAPL1x8/AM18x_and_OMAP1x7/AM17x

Signed-off-by: David Lechner <david@lechnology.com>