irqchip/irq-pruss-intc: add support for extra PRUSSEVTSEL interrupt cell
This adds support for an optional 4th interrupt cell needed for TI
AM18XX/OMAP-L139 SoCs. The event numbers on these SoCs could be one
of two events depending on the setting of PRUSSEVTSEL in the CFGCHIP3
syscon register. Therefore, an extra cell is needed to fully describe
the interrupt in this case.

The driver handles this by checking #interrupt-cells. If it is 4, then
it requires access to the ti,da830-cfgchip syscon registers for access
to PRUSSEVTSEL. Reference counting is used to provide first-come-first-
serve access to selecting the PRUSSEVTSEL value similar to how events
and channels are selected. The first user that needs to set PRUSSEVTSEL
gets to chose the value. Any additional users must request the same
value or will receive -EBUSY.

[1]: https://processors.wiki.ti.com/index.php/PRU_Subsystem_Differences_Between_OMAPL1x8/AM18x_and_OMAP1x7/AM17x

Signed-off-by: David Lechner <david@lechnology.com>