remoteproc: pru: future-proof PRU ID matching
Currently, to determine the ID (0 or 1) of a PRU core, the last 19 bits
of the physical address of the cores IRAM are compared to known values.
However, the PRUs on TI AM18XX have IRAM at 0x01c38000 and 0x01c3c000
respectively. The former conflicts with PRU1_IRAM_ADDR_MASK which could
cause PRU0 to be detected as PRU1. (The latter also conflicts with
TX_PRU1_IRAM_ADDR_MASK but it would still be correctly detected as
PRU1.)

This fixes the problem by moving the address matching offset values to
the device-specific data. This way the compatible string does half of
the work of narrowing down the addresses to two possibilities instead
of checking the address against all possible PRU types. This also lets
us narrow down the scope of the match from 19 bits to 16 bits.

After this, the TIAM18XX PRUs can be added without running into the
problems stated above.

Signed-off-by: David Lechner <david@lechnology.com>